Support up to 500 laps (room/race path)
Widen the lap count from byte (max 255) to int across the room, race, and
track-package paths so single race, time trial, and multiplayer rooms can be
set up to 500 laps. The byte cap was only an artifact of the old 16-lap limit,
not a deliberate design constraint.

Changes are a near 1:1 type widening:
  - byte -> int on every Laps/NrOfLaps storage and wire field (packets,
    GameRoom, TrackData, room snapshot/event, track-package manifest).
  - Serialization read/write switched to Int32 with matching payload sizes
    (client and server sides kept symmetric).
  - Lap limits/validation raised 16 -> 500 (session MaxLaps consts, room
    option list, settings clamp, server validation, menu range).

Protocol shape tests updated for the wider Laps field and the packet-shape
snapshot regenerated (RoomEvent Laps now 4 bytes). Full test suite passes
(426/426). Cross-build note: this changes the wire format, so new and old
builds are not interoperable on lap-bearing packets.

+++ top_speed_net/TopSpeed.Server/Network/Model/GameRoom.cs
@@ -52,7 +52,7 @@ namespace TopSpeed.Server.Network
         public TrackData? TrackData { get; set; }
         public string TrackName { get; set; }
         public TrackPackageRef TrackSelection { get; set; }
-        public byte Laps { get; set; }
+        public int Laps { get; set; }
         public uint GameRulesFlags { get; set; }
         public HashSet<uint> TrackReadyPlayers { get; } = new HashSet<uint>();
         public HashSet<uint> ActiveRaceParticipantIds { get; } = new HashSet<uint>();
