+++ top_speed_net/TopSpeed.Shared/Data/Tracks/Types.cs
@@ -148,14 +148,14 @@ namespace TopSpeed.Data
         public IReadOnlyDictionary<string, TrackSoundSourceDefinition> SoundSources { get; }
         public string? SourcePath { get; }
         public int Length => Definitions.Length;
-        public byte Laps { get; set; }
+        public int Laps { get; set; }
 
         public TrackData(
             bool userDefined,
             TrackWeather weather,
             TrackAmbience ambience,
             TrackDefinition[] definitions,
-            byte laps = 0,
+            int laps = 0,
             string? name = null,
             string? version = null,
             IReadOnlyDictionary<string, string>? metadata = null,
@@ -187,7 +187,7 @@ namespace TopSpeed.Data
             IReadOnlyDictionary<string, TrackWeatherProfile>? weatherProfiles,
             TrackAmbience ambience,
             TrackDefinition[] definitions,
-            byte laps = 0,
+            int laps = 0,
             string? name = null,
             string? version = null,
             IReadOnlyDictionary<string, string>? metadata = null,
@@ -235,7 +235,7 @@ namespace TopSpeed.Data
                 : TrackWeatherProfile.CreatePreset(TrackWeatherProfile.DefaultProfileId, TrackWeather.Sunny);
         }
 
-        public TrackData WithLaps(byte laps)
+        public TrackData WithLaps(int laps)
         {
             return new TrackData(
                 UserDefined,
